backend/data_ingestion/requirements.txt
# Copyright 2025, Battelle Energy Alliance, LLC, ALL RIGHTS RESERVED

-e ../kg
pandas==2.2.3
PyMuPDF==1.24.14
python-docx==1.1.2
minio==7.2.14